Moved rascsi/rasctl/scsimon/rasdump.cpp to classes (for testability)
